host: for the last bytes/period, send read notification later
This read pointer indicate how much we have read, we can use it
to inform host side to refill the free periods.

But as it doesn't really mean that we have rendered all of them
(there may be several periods in the pipeline buffer), and for
the last bytes/period, host side will trigger stop immediately
once it got this notification.

To avoid lost for last several periods issue, we postpone this
notification to the finish of last bytes/period copy in dai,
which will be add in the following patches.

Signed-off-by: Keyon Jie <yang.jie@linux.intel.com>