allocator: simplify heap modification and allocation part 1
The allocator currently has different methods to allocate heap memory
depending on memory size requested. The heap also has to be defined
in two places atm i.e in memory.h and baytrail.x

This patch aligns the heap macros with the linker script and introduces
new flags to specify the allocation type.

Signed-off-by: Liam Girdwood <liam.r.girdwood@linux.intel.com>