core: Add support for pipeline creation and scheduling via IPC.
This is a large tightly coupled patch that adds support for pipeline
definition and scheduling configuration via host IPC.

The intention is that hosts can create and configure pipelines dynamically
at runtime based on use case. This involves defining a pipeline topologies
and scheduling configuration. It also includes configuration of
individual components within a pipeline.

Signed-off-by: Liam Girdwood <liam.r.girdwood@linux.intel.com>