SRC: Remove obsolete conversion filters
These conversion filters are no more used and will be replaced later with
new. The deletion is because the stop-band spec was changed from 0.51xFs
to 0.50xFs (no intentional aliasing of transition band) so the file
names of new default high-quality filters will be with different filename.
The filter spec is appended to file name to differentiate different
bandwidth filters for same rates fraction.

Signed-off-by: Seppo Ingalsuo <seppo.ingalsuo@linux.intel.com>