Testing: Add enums for file I/O based test components
This patch adds to ipc.h enabling for two new components to read and
write test vectors via file I/O. They are specific to test bench usage
on development computer only and are not real SOF components. This update
avoids need to edit this file for test bench build.

Signed-off-by: Seppo Ingalsuo <seppo.ingalsuo@linux.intel.com>