component: data: Make component get/set data host API generic
Make the host API and IPC to set and get component runtime data and values
generic with no bespoke commands at the IPC level.

Create a generic structure that descibes the type and size of data and
whether that data is appended to the IPC message or is available to
be DMAed from host.

Update the components to use the new structure. Some components like
SRC, tone and tone still require some work to export thier ABI.

Signed-off-by: Liam Girdwood <liam.r.girdwood@linux.intel.com>