ipc: mailbox: simplify mailbox usage.
There can be some confusion over mailbox naming wrt to in/out and up/down
naming with respect to DSP or host context.

Rename the mailboxes to imply either host initiated or DSP initiated IPC
so that it's obvious to driver and FW engineers which is the correct
mailbox to use.

This change also simplifies the IPC command replies by returning 1 if
the command function creates the reply message otherwise a generic
reply is used.

Signed-off-by: Liam Girdwood <liam.r.girdwood@linux.intel.com>