scheduler: Add support for 64bit timer driven scheduling.
Currently the scheduler will schedule task based on DAI DMA interrupts.
This patch also adds the option to also schedule based on a 64 bit timer.

The scheduler will now check the tasks in the queue and it will only
run the tasks in the current window otherwise it will set a timer to
call schedule() on the next task start time (in another window).

Signed-off-by: Liam Girdwood <liam.r.girdwood@linux.intel.com>