pipeline: preloader: simplify preloader by reusing pipeline copy()
Currently the preloader is bespoke and runs in the same context as
the IPC. This means some components may delay IPC and block other
components.

Reuse the existing pipeline copy() infrastructure so that the
preload is done after prepare() but before trigger() using the
same pipeline copy code on the same DSP core and context. i.e.
we preload a host buffer (for playback) and then schedule the
pipeline copy(s) after the preload IPC is done.

Signed-off-by: Liam Girdwood <liam.r.girdwood@linux.intel.com>