dai: xrun: dont stop and restart DAI on XRUN recovery
In most situations the DAI should not be stopped and the restarted on XRUNs,
e.g. upstream/downstream component XRUNs. This patch will keep the DAI
running in those circumstances.

There will be a subsequent patch that will allow for stopping and
restarting the DAI for XRUN's e.g. SSP XRUNs

Signed-off-by: Liam Girdwood <liam.r.girdwood@linux.intel.com>