topology: test: Refine name for test generator.
We have 5 critical arguments in the test generator but only used 3 for
the naming, this will make some files be overwritten.
Refine the name method to let these 5 arguments show in the tplg name.
Also we decide the pipe type in one of the argument, test-ssp.m4 and
test-src-ssp.m4 are almost the same, just remove one.

Signed-off-by: Pan Xiuli <xiuli.pan@linux.intel.com>