trace: dont resend old buffers if trace buffer overflows
The circular trace buffer can overflow if too much trace data is sent
before it can schedule a DMA copy to flush the data. This overflow
causes the avail bytes counter to be greater than the buffer size and
means the DMA copy is rescheduled until avail is 0 (repeating old data).

Fix this by resetting avail to 0 when we overflow.

Signed-off-by: Liam Girdwood <liam.r.girdwood@linux.intel.com>