rimage: add check for .bss section
We may have heap buffers for base FW, but they are handled by the
runtime function. We did not need to set them as bss in FW manifest.
So add check to mark the .bss segment as BSS and other segment as HEAP.
Then we only update manifest BSS segmet with BSS section.

Signed-off-by: Pan Xiuli <xiuli.pan@linux.intel.com>