dma-trace: Fix reschedule bug to avoid dma buffer overflow
We have wrong logic in reschedule, we always reschedule the trace_work
once the buffer is half full and new trace coming. We will delay the old
schedule before the old scheduled trace_work is finally run.
Thus the trace_work will like the carrot in front of the DSP as the donkey,
the DSP will never run the trace_work that scheduled in the furture
while we are in busy state and lots of trace are coming continuously.

Signed-off-by: Pan Xiuli <xiuli.pan@linux.intel.com>
Reviewed-by: Zhigang Wu <zhigang.wu@linux.intel.com>
Reviewed-by: Yan Wang <yan.wang@linux.intel.com>
Tested-by: Zhigang Wu <zhigang.wu@linux.intel.com>