mailbox: move multi-line macros to static inlines
Detected with Coverity:

In do_notify: The indentation of this code suggests it is nested when
it is not. (CWE-483)

multi_stmt_macro: The macro on this line expands into multiple
statements, only the first of which is nested within the preceding
parent while the rest are not.

	if (msg->rx_size && msg->rx_size < SOF_IPC_MSG_MAX_SIZE)
		mailbox_dspbox_read(msg->rx_data, 0, msg->rx_size);

Move mailbox macros to inline functions to remove the issue, keep
indentation the same and add typecasts as needed

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>