SOFT: Improve audio test synchronization code
The audio tests signals contain a start and end marker to extract the
test output to analyze from in-between the markers.

This patch fixes a fail in SRC tests that was caused by incorrect
synchronization to output data. The chirp upper frequency is limited
to 0.99 x Nyquist frequency since the Nyquist frequency in the end
aliased and caused a visible artefact to spectral plot and it triggered
a resampler filter impulse response that had some correlation with
start/end marker chirps. The cosine chirp initialization to start from
90 degrees (zero PCM code instead of max.) also cleaned the start impulse
and improved the look of spectrogram plot.

The syncronization was limited to seek from one measured channel only. The
sync finder also didn't handle properly multi-channel input.

The output file start marker scan length was reduced to be
idle+marker length. And the scan time for end was reduced to be
2xidle+marker length

Signed-off-by: Seppo Ingalsuo <seppo.ingalsuo@linux.intel.com>