platform: the capture pipeline should be scheduled in resume.
the issue case:
playback and capture work at same time. when pause and resume
the capture, the DSP will enter panic.
the issue analysis:
when pause the capture pipeline, the capture's last dma interrupt
happens, and the capture will be paused. at this time, each buffer
in capture pipeline will have the data unprocessed, sometimes it
might be full. if we did not trigger the capture pipeline in resume
process, the coming dma interrupt will have the xrun in the buffer
pointer update. it is the high possibility.

Signed-off-by: Wu Zhigang <zhigang.wu@linux.intel.com>