[RFC]platform:bug fix: alsa xrun cause fw panic.
when START command comes, the pipeline has to be prepared.
This case has to be taken into account:
when ALSA has XRUN, it will send stop/start commands to FW
and try to recover the pipepine. If we did not prepare
the pipeline before start, the recover process will fail.

Signed-off-by: Wu Zhigang <zhigang.wu@linux.intel.com>