Volume: Mitigate zipper noise by adjusting gain ramp parameters
Fixes: #103 "volume: zipper noise with fast changes".

The volume ramp length is increased from 64 ms to 250 ms. The vol_work
rate to update gain value is increased from every 2 ms to 1 ms. The
volume ramp macros are redefined to enter directly the desired ramp
length in milliseconds to ease the tuning.

Also the volume gain Qxy notation (currently Q2.16) is defined with macros
to avoid magic numbers in code in ramp computation.

Signed-off-by: Seppo Ingalsuo <seppo.ingalsuo@linux.intel.com>