Mixer: Remove attenuation and apply saturation to mixed audio
This patch removes the down scaling of inputs to preserve the signal level
non-modified in mixing. The 64 bit mix is saturated (clamp) to prevent
overflow.

It is recommended to instead use in topology volume components with safe
gain set in the inputs of mixer (-6/-9.5/-12 dB for 2/3/4 inputs) if
saturation in mixing must be prevented.

Signed-off-by: Seppo Ingalsuo <seppo.ingalsuo@linux.intel.com>