EQ FIR: Improve robustness with configuration blobs
This patch prevents an error if the EQ responses to channels mapping
table in the configuration blob contains less channels than current
number of channels in firmware. Without this check the lookup from blob
can go past the table. Reporting an error depended on successive blob
content to detect corrupt blob and was not guaranteed.

The situation is not changed to stop to error but instead extrapolate
the table by applying 1st channel EQ for additional channels. It e.g.
helps to duplicate mono effect EQ to all channels without need to make
blobs with mapping to match max. channels count of SOF. The used response
index for each channel can be seen from trace.

Signed-off-by: Seppo Ingalsuo <seppo.ingalsuo@linux.intel.com>