EQ IIR: Update IPC to ALSA convention, fix channel bypass, and do cleanup
This patch updates binary IPC to similar style as done previously for FIR
equalizer. The blob size needs to be added as first word in the blob.

The internals of IIR coefficients format is exposed to uapi/eq.h level.
Also the blob parsing code is changed to use the structures instead of
hardcoded word offsets.

The EQ assign or switch to bypass (via -1) is fixed so there is no more
need to define a bypass EQ shape into the blob.

Finally C code style is updated to pass the commit checker scripts.

Signed-off-by: Seppo Ingalsuo <seppo.ingalsuo@linux.intel.com>