dw-dma: refactor interrupt registration and unregistration
Changes flow of dw-dma interrupt registration and unregistration:
Now it is possible for slave cores to register for dw-dma handler.
Also registration happens before start and not on FW load
during probing. This way we are not wasting runtime heap memory
and also it will allow us to dynamically change execution core
for certain pipelines.

Signed-off-by: Tomasz Lauda <tomasz.lauda@linux.intel.com>