component: add cache ops
Adds cache operation for every component.
This method allows performing cache operation on
current execution core based on selected cache command.
It's required in order to allow pipeline task switching to
slave cores. Master core will handle IPC to create and
configure pipeline, but later the whole pipeline with
components needs to be flushed to SRAM for other cores
to properly read it.

Signed-off-by: Tomasz Lauda <tomasz.lauda@linux.intel.com>