EQ IIR: Change initialization and simplify dcache invalididate function
This patch changes equalizer initialization to store the allocated
delay lines buffers address in a single pointer and size component
data. This simplifies the dcache invalidate code avoids knowledge of
filter internals that may change for future intrinsic optimized
version.

The EQ reconfigure and responses switching is by this patch changed to
happen in prepare() to avoid doing unnecessary allocation and free
operations in typical component life cycle. Prior to to prepare sent
configuration is just stored but not initialized into equalizers.
On-the-fly response switching can be enabled later with some work for
the component.

Signed-off-by: Seppo Ingalsuo <seppo.ingalsuo@linux.intel.com>