IPC: Increase max size from 128 to 384
This patch increases IPC length to be sufficient for e.g. 2x 12th order
IIR configuration. This is near to minimum feasible to enable testing of
equalizers. There may be need to adjust the limit up even more later but
this is the proposal now until we know better the requirements.

Signed-off-by: Seppo Ingalsuo <seppo.ingalsuo@linux.intel.com>