Enforce implicit falltrough check for rimage and host builds
This patch adds the -Wimplicit-fallthrough=3 option into Makefiles to
check the switch case statements for possibly missed break statements.
Hence the build for those targets will need at least GCC7.

The only fix needed is for rimage target where a comment for falltrough
is added to mark it as intentional. Note that increase of check level
to 4 would intercept also such comments so it is not yet done. Other
modules like dai use already the commenting for intended falltrough.

Signed-off-by: Seppo Ingalsuo <seppo.ingalsuo@linux.intel.com>