EQ IIR: Prevent reconfigure during playback and add traces
This patch prevents the not supported reconfiguration during playback or
capture. The request is not returned as error because then the kernel
driver will automatically apply the binary command in next idle and the
new response is configured and activated.

Some traces are added to help debugging the driver. The configuration
size is traced and and trace code "eis" is added if the received
configuration has zero length.

Signed-off-by: Seppo Ingalsuo <seppo.ingalsuo@linux.intel.com>