task: clear interrupt earlier
list_item_for_safe is safe for item deletion, but when a new item is
appended to the list. It seems the item could not be go through.
To fix this limitation, move interrupt clear to make sure every IRQ is
handled.
Do not try to handle two task with one IRQ handler.

Signed-off-by: Pan Xiuli <xiuli.pan@linux.intel.com>