EQ IIR: Add support for S16_LE and S24_4LE data
This patch adds processing functions for S16_LE and S24_4LE data.
Existing support was only S32_LE. The lower word length filters are
executed with 32 bit algorithm core with left shifted input to keep the
data in MSB side. The output is shifted, rounded and saturated to the
lower precision.

Signed-off-by: Seppo Ingalsuo <seppo.ingalsuo@linux.intel.com>