eq: iir: don't destroy eq config in reset
Remove freeing of config data in reset. If pm is
not enabled, stopping the stream causes component
reset in firmware. When you start a new stream the
eq parameters are gone and defaults will be used.
When pm is enabled this is not a problem because
component is recreated from swidget in kernel side.

Signed-off-by: Jaska Uimonen <jaska.uimonen@intel.com>