SRC: Use 32 bit filter coefficients in host testbench
This patch adds definition of CONFIG_HOST for SOF host version build
and selects for SRC the better quality and full conversions set in
such case.

The problem with default for gcc was that the conversions set for
non-optimized xtensa gcc is very limited and the quality has been
lowered to save resources. Hence the testbench creates huge amount
of fails without this change. With this patch the SRC test tracks the
quality of coefficients for optimized HiFiEP/HiFi3 version code.

Signed-off-by: Seppo Ingalsuo <seppo.ingalsuo@linux.intel.com>