debugability: Fix potential buffer overflow in mbox traces
Buffer overflow was possible when mbox traces of varying length were
logged consecutively, near end of the buffer, petentially leading to
next segment of memory being overwritten with garbage.
Also log entry was pasted twice, once to the on-stack buffer, and then
separately to the mailbox. Now data is being copied to both from the same
source.
Also separating this to a function call, reduces overall .text section of FW.

Signed-off-by: ArturX Kloniecki <arturx.kloniecki@linux.intel.com>