Sue Creek: platform support, including IPC over SPI
This patch adds support for the Sue Creek platform, including IPC
over SPI. With it the platform has been tested to boot successfully, to
send a "firmware ready" IPC to the host, to receive a trace
configuration IPC from the host and to reply to it. No audio
functionality has been tested so far, so it can still turn out, that
run-time IPC isn't working properly yet.

Signed-off-by: Guennadi Liakhovetski <guennadi.liakhovetski@linux.intel.com>