EQ: Fix FIR and IIR equalizers for the new user ABI
This patch contains the bug fixes for the EQ components to make them
work correctly. The topology and tools update is addressed separately. The
matching kernel patch is needed as well.

The first fix is for uapi header file that had left the macros
not updated. It caused the setup blob parsing to fail.

The FIR setup traces are updated to be easy to read with new trace. The
functional change is in get data where the configuration data size
and ABI version are written back to response.

The IIR is changed similarly but in addition the hard coded count
increase is replaced by the macro calculated value from ABI header
file.

Signed-off-by: Seppo Ingalsuo <seppo.ingalsuo@linux.intel.com>